+++ plugins/woocommerce-admin/packages/admin-e2e-tests/src/pages/AnalyticsOverview.ts
@@ -68,7 +68,7 @@ export class AnalyticsOverview extends Analytics {
 				'.woocommerce-ellipsis-menu .woocommerce-ellipsis-menu__toggle'
 			);
 			await ellipsisMenu?.click();
-			await page.waitFor(
+			await page.waitForFunction(
 				() =>
 					! document.querySelector(
 						'.woocommerce-ellipsis-menu div[role=menu]'