+++ plugins/woocommerce-admin/packages/admin-e2e-tests/src/pages/WcHomescreen.ts
@@ -11,6 +11,7 @@ import {
 	getElementByAttributeAndValue,
 	waitForElementByTextWithoutThrow,
 	getElementByText,
+	waitForTimeout,
 } from '../utils/actions';
 import { BasePage } from './BasePage';
 
@@ -32,11 +33,11 @@ export class WcHomescreen extends BasePage {
 
 		if ( modal ) {
 			await this.clickButtonWithText( 'Next' );
-			await this.page.waitFor( 1000 );
+			await waitForTimeout( 1000 );
 			await this.clickButtonWithText( 'Next' );
-			await this.page.waitFor( 1000 );
+			await waitForTimeout( 1000 );
 			await this.click( '.components-guide__finish-button' );
-			await this.page.waitFor( 500 );
+			await waitForTimeout( 500 );
 		}
 	}
 
@@ -88,7 +89,7 @@ export class WcHomescreen extends BasePage {
 		await waitForElementByText( 'button', 'Hide this' );
 		const hideThisButton = await getElementByText( 'button', 'Hide this' );
 		await hideThisButton?.click();
-		await this.page.waitFor( 500 );
+		await waitForTimeout( 500 );
 	}
 
 	async waitForNotesRequestToBeLoaded(): Promise< void > {